requirements.txt
matplotlib
numpy
pandas >= 0.24.0
requests
tqdm
sklearn
xlrd >= 1.0.0
gitpython
